style: move interpreter stuff in its own file
